requirements.txt
# Flask Web Framework
flask==2.3.2
flask-cors==4.0.0

# ML and numerical packages
tensorflow-cpu==2.20.0
numpy==1.26.4

# Network requests for model download / debugging
requests==2.31.0

# Production WSGI server
gunicorn==21.2.0
